Cull and refactor Codex contract tests
