+++ packages/pi-codex-conversion/tests/adapter-tool-activation.test.ts
@@ -2,9 +2,8 @@ import test from "node:test";
 import assert from "node:assert/strict";
 import { DEFAULT_CODEX_CONVERSION_CONFIG } from "../src/adapter/activation/config.ts";
 import { syncAdapter } from "../src/adapter/activation/activation.ts";
-import { isAdapterRuntime, resolveCodexRuntimePlan } from "../src/adapter/activation/runtime-plan.ts";
+import { resolveCodexRuntimePlan } from "../src/adapter/activation/runtime-plan.ts";
 import type { AdapterState } from "../src/adapter/activation/state.ts";
-import { registerCodexTools } from "../src/extension/tools.ts";
 import { createCodexTurnState } from "../src/providers/openai-codex/turn-state.ts";
 
 function createToolHarness(activeTools: string[]) {
@@ -95,41 +94,4 @@ test("native Responses compaction stays scoped to OpenAI Codex and explicit prov
 	assert.equal(resolveCodexRuntimePlan(createContext({ provider: "openai", api: "openai-responses", id: "gpt-5" }) as never, config).nativeCompaction, false);
 	assert.equal(resolveCodexRuntimePlan(createContext({ provider: "openai-codex", api: "openai-codex-responses", id: "gpt-5" }) as never, config).nativeCompaction, true);
 	assert.equal(resolveCodexRuntimePlan(createContext({ provider: "my-provider", api: "openai-codex-responses", id: "gpt-5" }) as never, config).nativeCompaction, true);
-});
-
-test("voice-only honors selected extras by provider scope without enabling the adapter", () => {
-	const codexModel = { provider: "openai-codex", api: "openai-codex-responses", id: "gpt-5.6-luna", input: ["text", "image"] };
-	const otherModel = { provider: "other", api: "openai-responses", id: "other-model", input: ["text"] };
-	const extraTools = ["apply_patch", "view_image", "web_run", "imagegen"];
-	const cases = [
-		{ name: "off outside Codex scope", scope: { allProviders: "off", additionalProviders: [] }, model: otherModel, expected: [] },
-		{ name: "off on Codex", scope: { allProviders: "off", additionalProviders: [] }, model: codexModel, expected: extraTools },
-		{ name: "extra tools only", scope: { allProviders: "extras", additionalProviders: [] }, model: otherModel, expected: extraTools },
-		{ name: "all providers", scope: { allProviders: "on", additionalProviders: [] }, model: otherModel, expected: extraTools },
-	] as const;
-
-	for (const item of cases) {
-		const state = createAdapterState({
-			voiceFeaturesOnly: true,
-			scope: { allProviders: item.scope.allProviders, additionalProviders: [...item.scope.additionalProviders] },
-			tools: { customRustBinariesDir: "", webRun: false, imageGeneration: false, applyPatchOnly: true, viewImageOnly: true, webRunOnly: true, imageGenerationOnly: true, viewImageFallback: true },
-			beta: { codeMode: true, responsesLite: true },
-			compaction: { responsesCompaction: true },
-		});
-		const pi = createToolHarness(["read", "bash", "edit", "write", "parallel"]);
-		registerCodexTools(pi as never, { state } as never);
-		const statuses: unknown[] = [];
-		const ctx = createContext(item.model, statuses);
-
-		syncAdapter(pi as never, ctx as never, state);
-
-		assert.deepEqual(pi.activeTools().filter((name) => extraTools.includes(name)), item.expected, item.name);
-		assert.ok(item.expected.every((name) => pi.registeredTools().has(name)), `${item.name}: selected extras registered`);
-		assert.ok(["read", "bash", "edit", "write", "parallel"].every((name) => pi.activeTools().includes(name)), `${item.name}: unrelated tools preserved`);
-		assert.ok(["exec_command", "write_stdin", "exec", "wait"].every((name) => !pi.activeTools().includes(name)), `${item.name}: adapter tools suppressed`);
-		const plan = resolveCodexRuntimePlan(ctx as never, state.config);
-		assert.equal(isAdapterRuntime(plan), false, item.name);
-		assert.equal(plan.nativeCompaction, false, item.name);
-		assert.equal(statuses.at(-1), undefined, `${item.name}: adapter status suppressed`);
-	}
 });
