+++ packages/pi-codex-conversion/tests/websocket-prewarm.test.ts
@@ -1,0 +1,77 @@
+import assert from "node:assert/strict";
+import test from "node:test";
+import { DEFAULT_CODEX_CONVERSION_CONFIG } from "../src/adapter/activation/config.ts";
+import { createCodexExtensionRuntime } from "../src/extension/runtime.ts";
+import { prewarmOpenAICodexWebSocket } from "../src/providers/openai-codex-custom-provider.ts";
+import { isWebSocketSseFallbackActive, recordWebSocketSseFallback } from "../src/providers/openai-codex/websocket.ts";
+import {
+	ScriptedWebSocket,
+	collectStream,
+	createRegisteredCodexProvider,
+	installScriptedWebSocket,
+	websocketSuccess,
+} from "./openai-codex-test-support.ts";
+import {
+	type ResponseCreateFrame,
+	context,
+	model,
+	sentFrames,
+	streamOptions,
+	unfinishedResponse,
+	user,
+} from "./websocket-test-support.ts";
+
+test("transport reset preserves session-sticky SSE until shutdown", () => {
+	const runtime = createCodexExtensionRuntime({ sendUserMessage: () => undefined } as never);
+	const sessionId = "sticky-reset";
+	recordWebSocketSseFallback(sessionId);
+	runtime.resetTransport(sessionId);
+	assert.equal(isWebSocketSseFallbackActive(sessionId), true);
+	runtime.shutdownTransport(sessionId);
+	assert.equal(isWebSocketSseFallbackActive(sessionId), false);
+});
+
+test("post-compaction transport reset restores WebSocket eligibility", () => {
+	const runtime = createCodexExtensionRuntime({ sendUserMessage: () => undefined } as never);
+	const sessionId = "sticky-compaction-reset";
+	recordWebSocketSseFallback(sessionId);
+	runtime.resetTransportAfterCompaction(sessionId);
+	assert.equal(isWebSocketSseFallbackActive(sessionId), false);
+});
+
+test("unfinished WebSocket prewarm cannot seed a continuation", async () => {
+	const restoreWebSocket = installScriptedWebSocket([
+		unfinishedResponse("resp_prewarm_pending", "queued"),
+		websocketSuccess,
+	]);
+	try {
+		const registered = createRegisteredCodexProvider({ codeMode: true });
+		const sessionId = "unfinished-prewarm";
+		const requestContext = context([user("same user", 1)]);
+		await assert.rejects(
+			prewarmOpenAICodexWebSocket(
+				model as never,
+				requestContext as never,
+				streamOptions(sessionId) as never,
+				{
+					getConfig: () => ({
+						openai: DEFAULT_CODEX_CONVERSION_CONFIG.openai,
+						beta: { ...DEFAULT_CODEX_CONVERSION_CONFIG.beta, codeMode: true },
+					}),
+					turnState: registered.turnState,
+				},
+			),
+		);
+
+		await collectStream(registered.provider.streamSimple(
+			model as never,
+			requestContext as never,
+			streamOptions(sessionId) as never,
+		));
+		assert.equal(ScriptedWebSocket.opened, 2);
+		assert.equal((sentFrames()[0] as ResponseCreateFrame & { generate?: boolean }).generate, false);
+		assert.equal(sentFrames()[1]?.previous_response_id, undefined);
+	} finally {
+		restoreWebSocket();
+	}
+});
